extract dataclasses for heat2018
